Used viewinflater to put form templates dynamically into second activity, transformed UniversityCourse into a Parcelable class so that objects can be passed from one activity to another.
Creating third summary output activity (probably a simple textview) and then passing object data to fourth activity which will show things on the calendar.

mprogress bar needs to be updated and exception throwing needs to be checked.

Google API only allows reading and writing data. Need code for a calendar display that provides a summary of study times that is preferably filterable.