Check error message and exceptions work well. The Parcelables pass data properly and the GUI also accepts data properly and can pass data successfully to activity 4. Also, kinks in calculate methods of UniversityCourse ironed out.
Calendar display logic built. Need to run and debug the logic and graphical user interface. Also need to add the import Google Calendar function and then maybe a splash page for startup and for ending.

mprogress bar needs to be updated and exception throwing needs to be checked.

Google API only allows reading and writing data. Need code for a calendar display that provides a summary of study times that is preferably filterable.