DEBUGGING.
Data passes successfully to YourCalendarActivity, where calendar generates and so do textviews. Attempting to use Gridlines in ConstraintLayout to organize textview positions for screens of different sizes.
Also, next and previous buttons work fine except for the previous button when it is clicked during the 2nd month (going back to the 1st month--what's wrong?).

Log messages added in YourCalendarActivity and following ImportGoogleCalendarActivity, so far data is passing as expected.
Need to check error messages and exceptions thrown properly and that GUI works as expected in YourCalendarActivity.

Logic is complete and needs debugging.

mprogress bar needs to be updated to newer non-deprecated structure and perhaps splash pages should be added.

ImportGoogleCalendarActivity would preferably utilize a new calendar inserted into the user's calendar list.