DEBUGGING.
Data passes successfully to YourCalendarActivity, where calendar generates and so do textviews and gridlines permit organization of data in space in combination with a boolean matrix.
Also, next and previous buttons work fine except for the previous button when it is clicked during the 2nd month (this is because the calendar updates from the 1st day of the new month to be viewed, and the first day may not be within the semester date range and as such no data containers exist for that day).

ImportGoogleCalendarActivity is stopping due to an error size = 99; index = 100;

Need to check error messages and exceptions thrown properly and that GUI works as expected in YourCalendarActivity.

Logic is complete and needs debugging.

mprogress bar needs to be updated to newer non-deprecated structure and perhaps splash pages should be added.

ImportGoogleCalendarActivity would preferably utilize a new calendar inserted into the user's calendar list.