DEBUGGING.
Data passes successfully to YourCalendarActivity, where calendar generates and so do textviews and gridlines permit organization of data in space in combination with a boolean matrix.
Also, next and previous buttons work fine. ImportGoogleCalendarActivity is stopping due to an error size = 99; index = 100 (to be determined what this means). All logic is complete and now debugging.

TODO:
0) Resolve the ImportGoogleCalendarActivity error, and create a proper layout for it
1) Need to check error messages and exceptions thrown properly and that GUI works as expected in YourCalendarActivity.
2) mprogress bar needs to be updated to newer non-deprecated structure and perhaps splash pages should be added.
3) YourCalendarActivity calendar_row may need to be retrofitted with scrollviews and the boolean matrix enlarged to accommodate more events in a given day
4) Formatting on yourcalendaractivity so that there are gridlines between days and rows and headings
5) ImportGoogleCalendarActivity would preferably utilize a new calendar inserted into the user's calendar list.